+++ feature-account-impl/src/main/java/jp/co/soramitsu/feature_account_impl/presentation/account/AccountsDiffCallback.kt
@@ -1,5 +1,6 @@
 package jp.co.soramitsu.feature_account_impl.presentation.account
 
+import jp.co.soramitsu.feature_account_impl.presentation.account.model.AccountModel
 import jp.co.soramitsu.feature_account_impl.presentation.common.groupedList.BaseGroupedDiffCallback
 import jp.co.soramitsu.feature_account_impl.presentation.view.advanced.network.model.NetworkModel
