+++ feature-account-impl/src/main/java/jp/co/soramitsu/feature_account_impl/presentation/account/mixin/api/AccountListingMixin.kt
@@ -1,8 +1,9 @@
-package jp.co.soramitsu.feature_account_impl.presentation.account
+package jp.co.soramitsu.feature_account_impl.presentation.account.mixin.api
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import io.reactivex.disposables.CompositeDisposable
+import jp.co.soramitsu.feature_account_impl.presentation.account.model.AccountModel
 
 class AccountListing(val groupedAccounts: List<Any>, val selectedAccount: AccountModel)
